разделил по файлам и настроил работу
