--- laba-14_05_23.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.33414.496
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "laba-14_05_23", "laba-14_05_23\laba-14_05_23.vcxproj", "{D8DDA2D5-4758-48D9-A0AF-ABE74DDCBAAC}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D8DDA2D5-4758-48D9-A0AF-ABE74DDCBAAC}.Debug|x64.ActiveCfg = Debug|x64
-		{D8DDA2D5-4758-48D9-A0AF-ABE74DDCBAAC}.Debug|x64.Build.0 = Debug|x64
-		{D8DDA2D5-4758-48D9-A0AF-ABE74DDCBAAC}.Debug|x86.ActiveCfg = Debug|Win32
-		{D8DDA2D5-4758-48D9-A0AF-ABE74DDCBAAC}.Debug|x86.Build.0 = Debug|Win32
-		{D8DDA2D5-4758-48D9-A0AF-ABE74DDCBAAC}.Release|x64.ActiveCfg = Release|x64
-		{D8DDA2D5-4758-48D9-A0AF-ABE74DDCBAAC}.Release|x64.Build.0 = Release|x64
-		{D8DDA2D5-4758-48D9-A0AF-ABE74DDCBAAC}.Release|x86.ActiveCfg = Release|Win32
-		{D8DDA2D5-4758-48D9-A0AF-ABE74DDCBAAC}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {9C040E27-15A5-40B1-AA42-067137EF160F}
-	EndGlobalSection
-EndGlobal
