--- laba-14_05_23/CTreeLib.cpp
@@ -1,4 +1,0 @@
-//#include "CTreeLib.h"
-
-//#include "CTreeLib.h"
-
