--- laba-14_05_23/x64/Debug/laba-14_05_23.tlog/laba-14_05_23.lastbuildstate
@@ -1,2 +1,0 @@
-PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.35.32215:TargetPlatformVersion=10.0.22000.0:
-Debug|x64|C:\Users\user\Desktop\Laba_Sem_2___14.05.23\|
