+++ testLaba_26.02.23/testLaba_26.02.23.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.33414.496
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "testLaba_26.02.23", "testLaba_26.02.23\testLaba_26.02.23.vcxproj", "{9877DCB9-9DC9-4FE6-9B65-670ABB84EB48}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{9877DCB9-9DC9-4FE6-9B65-670ABB84EB48}.Debug|x64.ActiveCfg = Debug|x64
+		{9877DCB9-9DC9-4FE6-9B65-670ABB84EB48}.Debug|x64.Build.0 = Debug|x64
+		{9877DCB9-9DC9-4FE6-9B65-670ABB84EB48}.Debug|x86.ActiveCfg = Debug|Win32
+		{9877DCB9-9DC9-4FE6-9B65-670ABB84EB48}.Debug|x86.Build.0 = Debug|Win32
+		{9877DCB9-9DC9-4FE6-9B65-670ABB84EB48}.Release|x64.ActiveCfg = Release|x64
+		{9877DCB9-9DC9-4FE6-9B65-670ABB84EB48}.Release|x64.Build.0 = Release|x64
+		{9877DCB9-9DC9-4FE6-9B65-670ABB84EB48}.Release|x86.ActiveCfg = Release|Win32
+		{9877DCB9-9DC9-4FE6-9B65-670ABB84EB48}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {83824D70-086A-4ECE-A592-E5366B48A9E2}
+	EndGlobalSection
+EndGlobal
