+++ testLaba_26.02.23/testLaba_26.02.23/testLaba_26.02.23.vcxproj
@@ -21,8 +21,8 @@
   <PropertyGroup Label="Globals">
     <VCProjectVersion>16.0</VCProjectVersion>
     <Keyword>Win32Proj</Keyword>
-    <ProjectGuid>{d8dda2d5-4758-48d9-a0af-abe74ddcbaac}</ProjectGuid>
-    <RootNamespace>laba140523</RootNamespace>
+    <ProjectGuid>{9877dcb9-9dc9-4fe6-9b65-670abb84eb48}</ProjectGuid>
+    <RootNamespace>testLaba260223</RootNamespace>
     <WindowsTargetPlatformVersion>10.0</WindowsTargetPlatformVersion>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
@@ -127,12 +127,14 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="CTreeLib.cpp" />
-    <ClCompile Include="laba-14_05_23.cpp" />
+    <ClCompile Include="CDataLib.cpp" />
+    <ClCompile Include="testLaba_26.02.23.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="CMass.h" />
+    <ClInclude Include="CProduct.h" />
     <ClInclude Include="CTreeLib.h" />
-    <ClInclude Include="CListLib.hpp" />
+    <ClInclude Include="CDataLib.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
